Move Builder and Chunker into own files
